+++ plugin_ide.ui/src-dltk-branch/_org/eclipse/dltk/ui/text/ScriptSourceViewerConfiguration.java
@@ -9,19 +9,14 @@
  *******************************************************************************/
 package _org.eclipse.dltk.ui.text;
 
-import static melnorme.utilbox.core.Assert.AssertNamespace.assertNotNull;
 import melnorme.lang.ide.core.LangNature;
 import melnorme.lang.ide.ui.text.AbstractLangSourceViewerConfiguration;
 
 import org.eclipse.cdt.ui.text.IColorManager;
-import org.eclipse.dltk.compiler.util.Util;
 import org.eclipse.dltk.internal.ui.text.HTMLTextPresenter;
 import org.eclipse.dltk.internal.ui.text.hover.EditorTextHoverDescriptor;
 import org.eclipse.dltk.internal.ui.text.hover.EditorTextHoverProxy;
-import org.eclipse.dltk.ui.CodeFormatterConstants;
 import org.eclipse.dltk.ui.DLTKUIPlugin;
-import org.eclipse.dltk.ui.text.util.AutoEditUtils;
-import org.eclipse.dltk.ui.text.util.TabStyle;
 import org.eclipse.jface.preference.IPreferenceStore;
 import org.eclipse.jface.text.DefaultInformationControl;
 import org.eclipse.jface.text.IDocument;
@@ -29,11 +24,9 @@ import org.eclipse.jface.text.IInformationControl;
 import org.eclipse.jface.text.IInformationControlCreator;
 import org.eclipse.jface.text.ITextHover;
 import org.eclipse.jface.text.ITextViewerExtension2;
-import org.eclipse.jface.text.formatter.IContentFormatter;
 import org.eclipse.jface.text.information.IInformationPresenter;
 import org.eclipse.jface.text.information.IInformationProvider;
 import org.eclipse.jface.text.information.InformationPresenter;
-import org.eclipse.jface.text.quickassist.IQuickAssistAssistant;
 import org.eclipse.jface.text.source.Annotation;
 import org.eclipse.jface.text.source.IAnnotationHover;
 import org.eclipse.jface.text.source.ISourceViewer;
@@ -46,27 +39,9 @@ import _org.eclipse.jdt.internal.ui.text.HTMLAnnotationHover;
 /* FIXME: DLTK review uses of other DLTK internal classes, possibly add them. */
 public abstract class ScriptSourceViewerConfiguration extends AbstractLangSourceViewerConfiguration {
 
-	private String fDocumentPartitioning;
-
 	public ScriptSourceViewerConfiguration(IPreferenceStore preferenceStore, IColorManager colorManager, 
-			AbstractDecoratedTextEditor editor, String partitioning) {
+			AbstractDecoratedTextEditor editor) {
 		super(preferenceStore, colorManager, editor);
-		
-		fDocumentPartitioning = assertNotNull(partitioning);
-
-		initializeScanners();
-	}
-
-	protected void initializeScanners() {
-	}
-	
-	@Override
-	public String getConfiguredDocumentPartitioning(ISourceViewer sourceViewer) {
-		return fDocumentPartitioning;
-	}
-	
-	protected IColorManager getColorManager() {
-		return getColorManager2();
 	}
 	
 	/* FIXME: DLTK: review text hovers */
@@ -166,75 +141,7 @@ public abstract class ScriptSourceViewerConfiguration extends AbstractLangSource
 			}
 		};
 	}
-
-	@Override
-	public IContentFormatter getContentFormatter(ISourceViewer sourceViewer) {
-		return super.getContentFormatter(sourceViewer);
-	}
-
-	/* FIXME: DLTK: review this code getIndentPrefixes */
-	@Override
-	public String[] getIndentPrefixes(ISourceViewer sourceViewer, String contentType) {
-		if (fPreferenceStore == null) {
-			return super.getIndentPrefixes(sourceViewer, contentType);
-		}
-		final TabStyle tabStyle = getTabStyle();
-		final int tabWidth = getTabWidth(sourceViewer);
-		final int indentWidth = getIndentationSize(sourceViewer);
-		if (tabStyle != TabStyle.TAB && indentWidth < tabWidth) {
-			return new String[] { AutoEditUtils.getNSpaces(indentWidth), "\t", Util.EMPTY_STRING };
-		} else if (tabStyle == TabStyle.TAB) {
-			return getIndentPrefixesForTab(tabWidth);
-		} else {
-			return getIndentPrefixesForSpaces(tabWidth);
-		}
-	}
-
-	protected TabStyle getTabStyle() {
-		if (fPreferenceStore != null) {
-			TabStyle tabStyle = TabStyle.forName(fPreferenceStore
-					.getString(CodeFormatterConstants.FORMATTER_TAB_CHAR));
-			if (tabStyle != null) {
-				return tabStyle;
-			}
-		}
-		return TabStyle.TAB;
-	}
-
-	@Override
-	public int getTabWidth(ISourceViewer sourceViewer) {
-		if (fPreferenceStore == null)
-			return super.getTabWidth(sourceViewer);
-		return fPreferenceStore.getInt(CodeFormatterConstants.FORMATTER_TAB_SIZE);
-	}
-
-	protected int getIndentationSize(ISourceViewer sourceViewer) {
-		if (fPreferenceStore == null)
-			return super.getTabWidth(sourceViewer);
-		return fPreferenceStore
-				.getInt(CodeFormatterConstants.FORMATTER_INDENTATION_SIZE);
-	}
-
-	/**
-	 * Computes and returns the indent prefixes for space indentation and the
-	 * given <code>tabWidth</code>.
-	 * 
-	 * @param tabWidth
-	 *            the display tab width
-	 * @return the indent prefixes
-	 * @see #getIndentPrefixes(ISourceViewer, String)
-	 */
-	protected String[] getIndentPrefixesForSpaces(int tabWidth) {
-		final String[] indentPrefixes = new String[tabWidth + 2];
-		indentPrefixes[0] = AutoEditUtils.getNSpaces(tabWidth);
-		for (int i = 0; i < tabWidth; i++) {
-			indentPrefixes[i + 1] = AutoEditUtils.getNSpaces(i) + '\t';
-		}
-		indentPrefixes[tabWidth + 1] = ""; //$NON-NLS-1$
-		return indentPrefixes;
-	}
-
-
+	
 	@Override
 	public IAnnotationHover getAnnotationHover(ISourceViewer sourceViewer) {
 		return new HTMLAnnotationHover(false) {
@@ -244,7 +151,7 @@ public abstract class ScriptSourceViewerConfiguration extends AbstractLangSource
 			}
 		};
 	}
-
+	
 	@Override
 	public IAnnotationHover getOverviewRulerAnnotationHover(ISourceViewer sourceViewer) {
 		return new HTMLAnnotationHover(true) {
@@ -253,11 +160,6 @@ public abstract class ScriptSourceViewerConfiguration extends AbstractLangSource
 				return isShowInOverviewRuler(annotation);
 			}
 		};
-	}
-
-	@Override
-	public IQuickAssistAssistant getQuickAssistAssistant(ISourceViewer sourceViewer) {
-		return super.getQuickAssistAssistant(sourceViewer);
 	}
 	
 }